fixed merge conclict issue
